Better move ancora reader to the new corpus package.
